+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # libsunspec changelog
 
-## v0.1.0 - 2021-07-21
+## v0.1.1 - 2023-07-21
+* remove register R/W delay
+* set default baud rate to 9600 for all examples
+
+## v0.1.0 - 2023-07-21
 * initial release